Fixes to Norwegian Bokmal locale (nb)
First of all, Norwegian uses ',' as a separator, not '.'. Thus, in
Norwegian you would write '3,2' not '3.2' as you do in English.

Secondly, in Norwegian both ' ' and '.' can be used as delimiter,
however, ' ' is primarily used.

Thirdly, according to the SI system the kilo- prefix has a lowercase k,
not uppercase K.