Clarify errors.messages.too_(long|short)
Previous messages literally read too_long: "is too long (not more than X characters)", too_short: "is too short (not less than X characters)" which IMO was confusing as it could be read as "is too short (i.e., is not less than…)" whereas the actual meaning is "is too short (it *should* not be less than…)".

The version I suggest translates to "is too long (more than X characters)" and "is too short (less than X characters)" respectively which is unambiguous. It can now be read as a natural sentence "Attribute IS too short (less than…)" without the user having to assume another verb for the details given in the parentheses.